+++ l10n_es_aeat/models/__init__.py
@@ -17,5 +17,6 @@
 ##############################################################################
 
 from . import aeat_report
+from . import aeat_report_tax_mapping
 from . import aeat_export_configuration
 from . import aeat_tax_code_mapping
