+++ l10n_es_aeat_mod111/views/mod111_view.xml
@@ -87,9 +87,6 @@
                     </group>
                 </group>
                 </group>
-                <group name="group_tax_lines" position="attributes">
-                    <attribute name="attrs">{'invisible': 1}</attribute>
-                </group>
             </field>
         </record>
