+++ l10n_es_aeat_mod216/views/mod216_view.xml
@@ -47,9 +47,6 @@
                         </group>
                     </group>
                 </group>
-                <group name="group_tax_lines" position="attributes">
-                    <attribute name="attrs">{'invisible': 1}</attribute>
-                </group>
             </field>
         </record>
