+++ l10n_es_aeat_mod296/views/mod296_view.xml
@@ -26,9 +26,6 @@
                         <field name="lines296" nolabel="1"/>
                     </group>
                 </group>
-                <group name="group_tax_lines" position="attributes">
-                    <attribute name="attrs">{'invisible': 1}</attribute>
-                </group>
             </field>
         </record>
