+++ l10n_es_aeat_mod303/__openerp__.py
@@ -26,7 +26,7 @@
 
 {
     "name": "AEAT modelo 303",
-    "version": "8.0.1.3.0",
+    "version": "8.0.1.4.0",
     'category': "Accounting & Finance",
     'author': "Guadaltech,"
               "AvanzOSC,"
