+++ l10n_es_aeat_mod303/models/mod303.py
@@ -1,25 +1,14 @@
-# -*- encoding: utf-8 -*-
-##############################################################################
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
-#    the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
-#    along with this program.  If not, see http://www.gnu.org/licenses/.
-#
-##############################################################################
+# -*- coding: utf-8 -*-
+# © 2013 - Guadaltech - Alberto Martín Cortada
+# © 2015 - AvanzOSC - Ainara Galdona
+# © 2014-2016 - Serv. Tecnol. Avanzados - Pedro M. Baeza
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
 from openerp import models, fields, api, _
 
 
 class L10nEsAeatMod303Report(models.Model):
-    _inherit = "l10n.es.aeat.report"
+    _inherit = "l10n.es.aeat.report.tax.mapping"
     _name = "l10n.es.aeat.mod303.report"
     _description = "AEAT 303 Report"
