+++ l10n_es_aeat_mod303/views/mod303_view.xml
@@ -78,7 +78,7 @@
                         </group>
                     </group>
                     <group class="oe_subtotal_footer oe_right"
-                           attrs="{'invisible': ['|', ('tax_lines', '=', []), ('state', '=', 'draft')]}"
+                           attrs="{'invisible': [('state', '=', 'draft')]}"
                            colspan="2">
                         <field name="resultado_liquidacion"
                                class="oe_subtotal_footer_separator"
@@ -94,6 +94,12 @@
                                domain="[('partner_id', '=', company_partner_id)]"
                                attrs="{'invisible': [('result_type', 'not in', ('B', 'I')), ('compensate', '=', False)]}"
                                 />
+                    </group>
+                    <group string="Tax lines"
+                           name="group_tax_lines"
+                           colspan="4"
+                           attrs="{'invisible': [('state', '=', 'draft')]}">
+                        <field name="tax_lines" nolabel="1"/>
                     </group>
                 </group>
             </field>
